+++ tests/e2e/multigpu/test_eval.py
@@ -7,8 +7,8 @@ from pathlib import Path
 
 import yaml
 from accelerate.test_utils import execute_subprocess_async
+from e2e.utils import check_tensorboard
 from transformers.testing_utils import get_torch_dist_unique_port
-from utils.tensorboard_ import check_tensorboard
 
 from axolotl.utils.dict import DictDefault
