+++ tests/e2e/patched/test_unsloth_qlora.py
@@ -6,7 +6,7 @@ import os
 from pathlib import Path
 
 import pytest
-from utils.tensorboard_ import check_tensorboard
+from e2e.utils import check_tensorboard
 
 from axolotl.cli import load_datasets
 from axolotl.common.cli import TrainerCliArgs
