+++ tests/e2e/test_embeddings_lr.py
@@ -7,7 +7,7 @@ import os
 import unittest
 from pathlib import Path
 
-from utils.tensorboard_ import check_tensorboard
+from e2e.utils import check_tensorboard
 
 from axolotl.cli import load_datasets
 from axolotl.common.cli import TrainerCliArgs
