+++ tests/e2e/test_packing_loss.py
@@ -6,8 +6,8 @@ import logging
 import os
 import unittest
 
+from e2e.utils import check_tensorboard
 from transformers.utils import is_torch_bf16_gpu_available
-from utils.tensorboard_ import check_tensorboard
 
 from axolotl.cli import load_datasets
 from axolotl.common.cli import TrainerCliArgs
